# Functions in session