--- a/cs/playground/FasterKVDiskReadBenchmark/Program.cs
+++ b/cs/playground/FasterKVDiskReadBenchmark/Program.cs
@@ -115,7 +115,7 @@ namespace FasterKVDiskReadBenchmark
                 Input input = default;
                 int i = 0;
 
-                var tasks = new (long, ValueTask<FasterKV<Key, Value, Input, Output, Empty>.ReadAsyncResult<MyFuncs>>)[batchSize];
+                var tasks = new (long, ValueTask<FasterKV<Key, Value>.ReadAsyncResult<Input, Output, Empty, MyFuncs>>)[batchSize];
                 while (true)
                 {
                     key = new Key(NumKeys * id + rand.Next(0, NumKeys));