--- a/cs/src/core/Allocator/BlittableAllocator.cs
+++ b/cs/src/core/Allocator/BlittableAllocator.cs
@@ -67,6 +67,11 @@ namespace FASTER.core
             return recordSize;
         }
 
+        public override int GetRecordSize<Input>(long physicalAddress, ref Input input)
+        {
+            return recordSize;
+        }
+
         public override int GetAverageRecordSize()
         {
             return recordSize;