--- a/cs/src/core/ClientSession/IClientSession.cs
+++ b/cs/src/core/ClientSession/IClientSession.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+#pragma warning disable 0162
+
+
+namespace FASTER.core
+{
+    internal interface IClientSession
+    {
+        void AtomicSwitch(int version);
+    }
+}