--- a/cs/src/core/Index/Interfaces/NullFasterSession.cs
+++ b/cs/src/core/Index/Interfaces/NullFasterSession.cs
@@ -1,0 +1,19 @@
+﻿namespace FASTER.core
+{
+    struct NullFasterSession : IFasterSession
+    {
+        public static readonly NullFasterSession Instance;
+
+        public void CheckpointCompletionCallback(string guid, CommitPoint commitPoint)
+        {
+        }
+
+        public void UnsafeResumeThread()
+        {
+        }
+
+        public void UnsafeSuspendThread()
+        {
+        }
+    }
+}