For https://github.com/microsoft/FASTER/issues/230.
Not sure about naming (`ISynchronizationListener`).
There is at least one breaking change (`ReadAsyncResult` is in different class [different number of type arguments]).
For now I duplicated state machine code for sync and async versions, if this is not ideal I can unify into one method again.
Needed to introduce `IClientSession`, but this seems to be one-time call so no harm there.