@marius-klimantavicius 2020-05-28
Fixed naming (at least to my liking).
Still missing tests and example.