@tli2 2020-06-02
@badrishc 

I only took a cursory look. In principle I think that's the right thing to do. For this particular case there also seems to be not too much actual code duplication.

If I am to be nit-picky I'd say we need a better model for when it is appropriate to use which, and reflect that in the API naming and documentation. We might also consider enforcing that separation in the code (e.g. if we decide that `OnThreadAsync` can only be called from sessions that support async, we can check that in FasterStateMachine.cs)