@marius-klimantavicius 2020-06-02
Should I experiment with moving `Input`, `Output` and `Context` out of `FasterKV` (to be part of method signature not of class itself, part of `IFunctions` and `ClientSession`)? Rationale - different callbacks might work with different `Intput`, `Output`, `Context`.