@marius-klimantavicius 2020-06-02
One (relatively annoying) issue with moving `Input`, `Output` and `Context` is that type inference for `NewSession` and `ResumeSession` do not work for primary `FasterKV<Key, Value>` (there is also legacy `FasterKV` for backwards compat).