@marius-klimantavicius 2020-06-06
Latest commit is for #272 - Added custom compaction functions. Note that in addition to having custom `IsDeleted` I added `Copy` and `CopyInPlace` to be able to deal with concurrency (now that I am writing this - do we really need these? Is it possible to compact *inside* mutable region? Probably yes as `untilAddress` can be tail).

Net462 is supported when using FASTER as binary dependency (e.g. via nuget) however I am embedding sources directly (in addition I sometimes do have more changes) thus locally I've removed `readonly` members, nullables, local static functions, etc. For PRs back to FASTER I am trying to keep it close to original but still some changes might come through (like extracting static local functions to full functions [this is more of a personal preference], I can move them back).