@marius-klimantavicius 2020-06-09
I think I will do one more change - I want to add an overload to `IVariableLengthStruct.GetLength` to take in `TInput`. It is possible that due to RMW I might want to change the size of the record based on the incoming `input`.