@marius-klimantavicius 2020-07-02
I see, I have only 16GB of RAM on this machine maybe it was due swapping (I might try the PerfTest branch a bit later).