@badrishc 2020-07-09
With a minor update to ClientSession.cs, I was able to restore performance to the same level as before. Please review if the change is ok with you.