@marius-klimantavicius 2020-07-10
Yep these are absolutely OK (I was hoping `new AsyncFasterSession(this)` to be optimized away but I was wrong). Having `readonly functions` is the same as before this PR (though once I figure out how to properly benchmark [provided my pc is powerful enough] I will check if perf is not affected when `functions` is `struct` [due to defensive copy though only matter for non-empty structs]).