@badrishc 2020-09-03
> One (relatively annoying) issue with moving `Input`, `Output` and `Context` is that type inference for `NewSession` and `ResumeSession` do not work for primary `FasterKV<Key, Value>` (there is also legacy `FasterKV` for backwards compat).
> 
> Can slightly reduce the noise by adding helper "builder" (still need to specify `Input`, `Output` and `Context` but functions can be inferred):
> 
> ```
> Builder<Input, Output, Context> For<Input, Output, Context>() => new Builder<...>(this);
> ```
> 
> ```
> using (var session = fasterKV.For<MyInput, MyOutput, MyContext>().NewSession(new MyFunctions()...)
> ```

@marius-klimantavicius: In PR https://github.com/microsoft/FASTER/pull/310, I have extended the builder slightly, to support writing of slightly cleaner code that produces a proper client session with the right generic params (i.e., specific `Function` instead of `IFunctions<>`):

```cs
store.For(new MyFunctions()).NewSession<MyFunctions>(...);
```

It is still ugly (`MyFunctions` has to be uttered twice), but at least user does not need to utter the types of `Input`, `Output`, `Context`.